Modified TestFinder to take an options object instead of positional arguments. It now takes 'testInProcess' and 'testOutOfProcess' as options. Also, its 'filter' option can now be either a string or a function--if it's a string, it's treated as a regular expression as before, but if it's a function, it's expected to take the filename of a test suite and return a boolean indicating whether or not the suite should be executed.
We do, however, need to update the Markdown documentation for the aforementioned modules to reflect these changes.

Also added test-e10s-porting.js, which runs all test suites that have been white-listed as being compatible w/ E10s in a separate process, even when '--e10s' isn't passed to cfx. This ensures that while we're porting the SDK to be compatible with e10s, we can still run the tests we know to work and ensure there aren't any regressions.