Added initial support for XULRunner+Gecko2.
Tested this out with 'cfx test -F trait -a xulrunner' when in the 'packages/jetpack-core' directory and it worked, but I haven't tried this out with 'cfx run' yet.

The chrome.manifest is specifically tweaked right now to only work when the application ID is xulapp@toolness.com, but we can change that later.

The documentation I consulted to make most of these changes is here:

https://developer.mozilla.org/en/XPCOM/XPCOM_changes_in_Gecko_1.9.3